# Conflicts:
#	src/literalizer/languages/systemverilog.py

+++ tests/integration/cases/list_mixed_int_widths/Go@v1_18.go
@@ -1,7 +1,7 @@
 package main
 
 func main() {
-my_data := []int{
+my_data := []int64{
 	1,
 	1099511627776,
 }
